Enemy AI fixed, Midi Fixed, Fixed GM, Added Quest Text
Enemy AI failed because they were listening to the wrong note (36 instead of 42) and because the midi manager was using the wrong midi (updated instead of together). The GM was also causing issues with the quest stuff because it had multiple instances of GameData, QuestData and Midi System on it. Finally, I added text that shoots up and fades updating the player on the quest. Added this for feedback.